+++ src/app/features/forms/common/utils/checkLowLeverageNotification/index.ts
@@ -1,7 +1,7 @@
 import { Draft } from '@reduxjs/toolkit';
 
 import { SliceState as LPFormSliceState } from '../../../lp-form/state';
-import { SliceState as SwapFormSliceState } from '../../../swap-form/reducer';
+import { SliceState as SwapFormSliceState } from '../../../swap-form/state';
 
 export const checkLowLeverageNotification = (
   state: Draft<SwapFormSliceState | LPFormSliceState>,
