+++ src/app/features/forms/swap-form/reducer.ts
@@ -1,11 +1,12 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { AMM, InfoPostSwapV1, Position } from '@voltz-protocol/v1-sdk';
+import { AMM, InfoPostSwapV1 } from '@voltz-protocol/v1-sdk';
 import { ContractReceipt } from 'ethers';
 
 import { getAmmProtocol } from '../../../../utilities/amm';
 import { stringToBigFloat } from '../../../../utilities/number';
 import { checkLowLeverageNotification, formLimitAndFormatNumber } from '../common/utils';
 import { pushLeverageChangeEvent } from './analytics';
+import { initialState } from './state';
 import {
   approveUnderlyingTokenThunk,
   confirmMarginUpdateThunk,
@@ -26,185 +27,6 @@ import {
   updateLeverageOptionsAfterGetPoolSwapInfo,
   validateUserInputAndUpdateSubmitButton,
 } from './utils';
-
-type ThunkStatus = 'idle' | 'pending' | 'success' | 'error';
-
-export type SliceState = {
-  submitButton: {
-    state:
-      | 'swap'
-      | 'margin-update'
-      | 'not-enough-balance'
-      | 'connect-wallet'
-      | 'approve'
-      | 'approving';
-    disabled: boolean;
-    message: {
-      text: string | null;
-      isError: boolean;
-    };
-  };
-  amm: AMM | null;
-  position: {
-    value: Position | null;
-    status: ThunkStatus;
-  };
-  walletBalance: {
-    value: number;
-    status: ThunkStatus;
-  };
-  walletTokenAllowance: {
-    value: number;
-    status: ThunkStatus;
-  };
-  // User-agnostic swap info about pool
-  poolSwapInfo: {
-    availableNotional: Record<'fixed' | 'variable', number>;
-    maxLeverage: Record<'fixed' | 'variable', number>;
-    status: ThunkStatus;
-  };
-  userInput: {
-    // Side chosen by user in the UI
-    mode: 'fixed' | 'variable';
-    // User-inputted notional amount
-    notionalAmount: {
-      value: number;
-      editMode: 'add' | 'remove';
-      error: string | null;
-    };
-    // User-inputted margin amount
-    marginAmount: {
-      value: number;
-      editMode: 'add' | 'remove';
-      error: string | null;
-    };
-    leverage: number | null;
-  };
-  // State of prospective swap
-  prospectiveSwap: {
-    // Leverage options
-    leverage: {
-      options: number[];
-      maxLeverage: string;
-    };
-    infoPostSwap: {
-      value: {
-        // Margin requirement for prospective swap
-        marginRequirement: number;
-        // Max margin that can be withdrawn after prospective swap
-        maxMarginWithdrawable: number;
-        // Average fixed rate obtained by the prospective swap
-        averageFixedRate: number;
-
-        // Token balance deltas resulted by prospective swap
-        fixedTokenDeltaBalance: number;
-        variableTokenDeltaBalance: number;
-        fixedTokenDeltaUnbalanced: number;
-
-        // Extra information about prospective swap
-        fee: number;
-        slippage: number;
-        gasFeeETH: number;
-      };
-      status: ThunkStatus;
-    };
-  };
-  swapConfirmationFlow: {
-    step: 'swapConfirmation' | 'waitingForSwapConfirmation' | 'swapCompleted' | null;
-    error: string | null;
-    txHash: string | null;
-  };
-  marginUpdateConfirmationFlow: {
-    step:
-      | 'marginUpdateConfirmation'
-      | 'waitingForMarginUpdateConfirmation'
-      | 'marginUpdateCompleted'
-      | null;
-    error: string | null;
-    txHash: string | null;
-  };
-  showLowLeverageNotification: boolean;
-};
-
-const initialState: SliceState = {
-  submitButton: {
-    state: 'connect-wallet',
-    disabled: true,
-    message: {
-      text: null,
-      isError: false,
-    },
-  },
-  amm: null,
-  position: {
-    value: null,
-    status: 'idle',
-  },
-  walletBalance: {
-    value: 0,
-    status: 'idle',
-  },
-  walletTokenAllowance: {
-    value: 0,
-    status: 'idle',
-  },
-  poolSwapInfo: {
-    availableNotional: {
-      fixed: 0,
-      variable: 0,
-    },
-    maxLeverage: {
-      fixed: 0,
-      variable: 0,
-    },
-    status: 'idle',
-  },
-  userInput: {
-    mode: 'fixed',
-    notionalAmount: {
-      value: 0,
-      editMode: 'add',
-      error: null,
-    },
-    marginAmount: {
-      value: 0,
-      editMode: 'add',
-      error: null,
-    },
-    leverage: null,
-  },
-  prospectiveSwap: {
-    leverage: {
-      options: [0, 0, 0],
-      maxLeverage: '--',
-    },
-    infoPostSwap: {
-      value: {
-        marginRequirement: 0,
-        maxMarginWithdrawable: 0,
-        averageFixedRate: 0,
-        fixedTokenDeltaBalance: 0,
-        variableTokenDeltaBalance: 0,
-        fixedTokenDeltaUnbalanced: 0,
-        fee: 0,
-        slippage: 0,
-        gasFeeETH: 0,
-      },
-      status: 'idle',
-    },
-  },
-  swapConfirmationFlow: {
-    step: null,
-    error: null,
-    txHash: null,
-  },
-  marginUpdateConfirmationFlow: {
-    step: null,
-    error: null,
-    txHash: null,
-  },
-  showLowLeverageNotification: false,
-};
 
 const slice = createSlice({
   name: 'swapForm',
