+++ src/app/features/forms/swap-form/utils/getAvailableMargin/index.ts
@@ -3,7 +3,7 @@ import { Position } from '@voltz-protocol/v1-sdk';
 
 import { stringToBigFloat } from '../../../../../../utilities/number';
 import { formLimitAndFormatNumber } from '../../../common/utils';
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 import { getProspectiveSwapNotional } from '../getProspectiveSwapNotional';
 import { hasExistingPosition } from '../hasExistingPosition';
