+++ src/app/features/forms/swap-form/utils/getEditPositionNotional/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 import { getEditPositionTokenBalance } from '../getEditPositionTokenBalance';
 
 export const getEditPositionNotional = (state: Draft<SliceState>) => {
