+++ src/app/features/forms/swap-form/utils/getEditPositionTokenBalance/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const getEditPositionTokenBalance = (state: Draft<SliceState>) => {
   let fixedTokenBalance = 0;
