+++ src/app/features/forms/swap-form/utils/getEditPositionVariableRate/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 import { getEditPositionNotional } from '../getEditPositionNotional';
 import { getVariableRate } from '../getVariableRate';
