+++ src/app/features/forms/swap-form/utils/getExistingPositionMode/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const getExistingPositionMode = (state: Draft<SliceState>) => {
   if (state.position.status !== 'success' || !state.position.value) {
