+++ src/app/features/forms/swap-form/utils/getExistingPositionNotional/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const getExistingPositionNotional = (state: Draft<SliceState>) => {
   if (state.position.status !== 'success' || !state.position.value) {
