+++ src/app/features/forms/swap-form/utils/getRealizedPnLFromFees/index.ts
@@ -1,7 +1,7 @@
 import { Draft } from '@reduxjs/toolkit';
 import { Position } from '@voltz-protocol/v1-sdk';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 import { hasExistingPosition } from '../hasExistingPosition';
 
 export const getRealizedPnLFromFees = (state: Draft<SliceState>) => {
