+++ src/app/features/forms/swap-form/utils/getVariableRate/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const getVariableRate = (state: Draft<SliceState>) => {
   return state.amm ? state.amm.variableApy : null;
