+++ src/app/features/forms/swap-form/utils/hasExistingPosition/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState } from '../../reducer';
+import { SliceState } from '../../state';
 
 export const hasExistingPosition = (state: Draft<SliceState>): boolean => {
   return state.position.status === 'success' && state.position.value !== null;
